% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{avg_runs}
\alias{avg_runs}
\title{Extract the average density of the last timesteps}
\usage{
avg_runs(run, n = 100)
}
\arguments{
\item{run}{a dataframe. It contains the variable time and the density of the
of the state variable.}

\item{n}{a integer. Define at how many timesteps from the last one should be
selected to average density}
}
\value{
a dataframe. It contains
}
\description{
This function extracts the density of the last timesteps and average it.
}
