% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{bind_scenar}
\alias{bind_scenar}
\title{Bind scenarii object together}
\usage{
bind_scenar(x, ...)
}
\arguments{
\item{l}{a list of scenarii object}

\item{var_name}{a character vector. Names of the parameters which are different
between scenarii}
}
\description{
Combine a list of scenarii in one
}
\details{
The scenario should describe one of init density values defined in
init_scenarii
}
\seealso{
init_scenarii
}
