% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_nbs}
\alias{check_nbs}
\title{Check the neighbor arguments}
\usage{
check_nbs(nbs)
}
\arguments{
\item{nbs}{string. Either "N", "P" or "NULL"}
}
\value{
stop if the argument is badly define
}
\description{
\code{indirect_facilitation_model} returns the simObj.
}
