% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_var}
\alias{check_var}
\title{Check the neighbor arguments}
\usage{
check_var(arg)
}
\arguments{
\item{arg}{vector List of arguments}
}
\value{
stop if the argument is badly defined
}
\description{
Check the neighbor arguments
}
