% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_z}
\alias{check_z}
\title{Check the neighbor arguments}
\usage{
check_z(z)
}
\arguments{
\item{z}{numeric. Either 4 or 8}
}
\value{
stop if the argument is badly defined
}
\description{
Check the neighbor arguments
}
