% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{clean_run}
\alias{clean_run}
\title{Clean averaged simulation runs}
\usage{
clean_run(x)
}
\arguments{
\item{run}{a tibble or a data.frame}
}
\value{
a tibble or a data.frame
}
\description{
Suppress simulation runs and spurious information
}
