% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{compute_as}
\alias{compute_as}
\title{Protection function for the protegee}
\usage{
compute_as(type = "linear")
}
\arguments{
\item{type}{}

\item{NE_proba}{}
}
\value{
a rate of transition
}
\description{
Protection function for the protegee
}
