% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{compute_light_paul_palette}
\alias{compute_light_paul_palette}
\title{Color blind palette}
\usage{
compute_light_paul_palette()
}
\value{
a named vector of 9 colors
}
\description{
Paul tol palette light qualitative, it can be found at 
https://personal.sron.nl/~pault/
}
