% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{compute_occurences}
\alias{compute_occurences}
\title{Compute the co-occurences between nurse and protegee}
\usage{
compute_occurences(x, ...)
}
\arguments{
\item{data}{a dataframe.}
}
\value{
a dataframe.
}
\description{
This function computes the density of cooccuring species.
It should be after avg_runs
}
