% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{compute_p_one_z}
\alias{compute_p_one_z}
\title{relationship between the effect of a nurse and the cost of defense}
\usage{
compute_p_one_z(gamma1, u)
}
\arguments{
\item{gamma1}{cost of defence}

\item{u}{shape of the curve.}
}
\value{
a rate of transition
}
\description{
relationship between the effect of a nurse and the cost of defense
}
