% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{compute_states}
\alias{compute_states}
\title{Compute the state result of a range of simulation}
\usage{
compute_states(x, ...)
}
\arguments{
\item{data}{a gradient object}

\item{type}{}

\item{possibles_states}{}
}
\description{
Compute the state result of a range of simulation
}
