% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{compute_tau}
\alias{compute_tau}
\title{Define the shape of the relationship between the protection effect and the
number of nurse}
\usage{
compute_tau(p = 9/10, z = 4)
}
\arguments{
\item{p}{the effect of one nurse}

\item{z}{the number of neighbors}
}
\value{
tau
}
\description{
Define the shape of the relationship between the protection effect and the
number of nurse
}
