% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{convert2scenarii}
\alias{convert2scenarii}
\title{Convert a gradient object to a scenarii one}
\usage{
convert2scenarii(x, scenario)
}
\arguments{
\item{x}{a gradient object}

\item{scenario}{a character vector of length one. See @details}
}
\description{
Convert a gradient object to a scenarii one
}
\details{
The scenario should describe one of init density values defined in
init_scenarii
}
\seealso{
init_scenarii
}
