% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{D_context}
\alias{D_context}
\title{Neighboring function}
\usage{
D_context(nbs, ND, PD, D, z)
}
\arguments{
\item{nbs}{}

\item{D}{}

\item{z}{}
}
\value{
definition of the probability
}
\description{
Define the probability of an event according to the Neighboring cells for a 
degraded cell.
}
