% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{def_state}
\alias{def_state}
\title{Define the state result of a simulation}
\usage{
def_state(nurse, protegee, sim_status, threshold = 10^-3)
}
\arguments{
\item{nurse}{a dbl of length 1. The density of the nurse at the steady
state.}

\item{protegee}{a dbl of length 1. The density of the protegee at the steady
state.}
}
\value{
a chr vector of length 1
}
\description{
This function gives the result of a simulation, i.e., "warning", "nurse",
"protegee", "coexistence"
}
