% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{define_double_state}
\alias{define_double_state}
\title{Define bistable states}
\usage{
define_double_state(scenar1, scenar2)
}
\arguments{
\item{scenar1}{a character string. It is the name the first scenario}

\item{scenar2}{a character string. It is the name the second scenario}
}
\description{
Define bistable states
}
