% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{degrade}
\alias{degrade}
\title{Degradation transition function}
\usage{
degrade(d)
}
\arguments{
\item{d}{rate of degradation.  1/d defines the degradation time expectancy.}
}
\value{
definition of the probability
}
\description{
Define the probability to be degraded for a empty cell.
}
