% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{df2list}
\alias{df2list}
\title{Convert a data.frame to a named list}
\usage{
df2list(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a list.
}
\description{
Convert a data.frame to a named list
}
\details{
Each element of the list contains a named vector. The names
are those of the columns of the input.
}
