% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{die}
\alias{die}
\title{Death transition function}
\usage{
die(m)
}
\arguments{
\item{m}{rate of mortality.  1/m defines the life expectancy.}
}
\value{
definition of the probability
}
\description{
Define the probability to die for a cell.
}
