% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{facilitate}
\alias{facilitate}
\title{Facilitation transition function}
\usage{
facilitate(ND, PD, D, z, f, nbs)
}
\arguments{
\item{f}{rate of local condition improvement.}
}
\value{
definition of the probability
}
\description{
Define the probability to be regenerated for a degraded cell thanks to the
improvement of local conditions by an adult plant.
}
