% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{filter.scenarii}
\alias{filter.scenarii}
\title{Subset runs}
\usage{
\method{filter}{scenarii}(data, ...)
}
\arguments{
\item{data}{a scenarii object}

\item{...}{filtering rules}
}
\value{
a scenarii object
}
\description{
Subset runs
}
\details{
https://cran.r-project.org/web/packages/dplyr/vignettes/programming.html
}
