% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_4_states_model.R
\name{four_states_sys}
\alias{four_states_sys}
\title{pair approximation system of ODEs with a nurse, a protege, empty and degraded states}
\usage{
four_states_sys(time, init, parms)
}
\arguments{
\item{...}{Nothing}
}
\value{
A list of ODEs
}
\description{
pair approximation system of ODEs with a nurse, a protege, empty and degraded states
}
