% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_transitions_functions.R
\name{fourneighbors}
\alias{fourneighbors}
\title{Compute number of neighbors out of 4 neighboring cells}
\usage{
fourneighbors(landscape, state = 1, bounds = 1)
}
\arguments{
\item{state}{A value, whose existence is checked within the
    neighbourhood of each cell.
  }

\item{bounds}{A vector with either one or four values specifying
    the type of boundaries, where 0 means open boundaries and 1 torus-like
    boundaries. The values are specified in the order bottom, left, 
    top, right.}
}
\value{
a matrix
}
\description{
This function is based on the simecol::neighbors function.
}
\details{
The performance of the function depends on the size of the matrices and the
  type of the boundaries, where open boundaries are faster than torus like
  boundaries. Function \code{\link{eightneighbours}} is even faster.
}
\seealso{
\code{\link{seedfill}}, \code{\link{eightneighbours}}, \code{\link{conway}}
}
