% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{identify_transition}
\alias{identify_transition}
\title{Identify abrupt transitions}
\usage{
identify_transition(data, threshold = 0.001)
}
\arguments{
\item{threshold}{numeric}

\item{v}{a vector}
}
\value{
factor
}
\description{
Identify abrupt transitions
}
\details{
It identifies changes in variable values and assign series
accordingly to groups.
}
\seealso{
plot_bifurcation
}
