% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_facilitation.R
\name{indirect_facilitation_model}
\alias{indirect_facilitation_model}
\title{Indirect facilitation model with a nurse and a protege}
\usage{
indirect_facilitation_model()
}
\arguments{
\item{...}{Nothing}
}
\value{
An ecological model of the class simObj with default definition.
}
\description{
\code{indirect_facilitation_model} returns the simObj.
}
\examples{
model <- indirect_facilitation_model() 

\dontrun{
 out <- sim(model)

}

}
