% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{init_scenarii}
\alias{init_scenarii}
\title{TODO: Document function
Initialize starting values of state variables}
\usage{
init_scenarii(type = "together", model = two_facilitation_model(),
  ini_cover = 0.8, low_cover = 0.05)
}
\arguments{
\item{type}{character. the scenario}

\item{ini_cover}{numeric initial cover for high starting cover}
}
\value{
A named list. Each element contains a numeric vector of initial
values of the state variables.
}
\description{
Compute initial values of states variable according to a scenario
}
\details{
when type = "all", all the scenarii are returned. High cover
scenarii set the total cover to ini_cover. It is divided by 2 in together. In
low_P or low_N scenarii, the cover of the rare species is of low_cover
divided by 2.
}
