% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{is_run_normal}
\alias{is_run_normal}
\title{Check integration problem in runs}
\usage{
is_run_normal(df)
}
\arguments{
\item{df}{a dataframe containing the run of an ODE model}
}
\value{
logical
}
\description{
Check if there is NaN (Not a Number) or negative densities during the run
}
