% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{Ncolonize}
\alias{Ncolonize}
\title{Colonization function for the nurse pair approximation}
\usage{
Ncolonize(N, NE, E, z, del, b, c, gamma1, nbs)
}
\arguments{
\item{N}{}

\item{NE}{}

\item{E}{}

\item{z}{}

\item{del}{}

\item{b}{}

\item{c}{}

\item{gamma1}{}

\item{nbs}{}
}
\value{
a rate of transition
}
\description{
Colonization function for the nurse pair approximation
}
