% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{NE_context}
\alias{NE_context}
\title{Neighboring function}
\usage{
NE_context(nbs, NE, E, z)
}
\arguments{
\item{nbs}{}
}
\value{
definition of the probabilityj
}
\description{
Neighboring function
}
