% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{Pcolonize}
\alias{Pcolonize}
\title{Colonization function for the nurse pair approximation}
\usage{
Pcolonize(P, N, NE, PE, E, z, del, b, c, g, n, nbs)
}
\arguments{
\item{nbs}{}
}
\value{
a rate of transition
}
\description{
Colonization function for the nurse pair approximation
}
