% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{PE_context}
\alias{PE_context}
\title{Neighboring function}
\usage{
PE_context(nbs, PE, E, z)
}
\arguments{
\item{nbs}{}

\item{E}{}

\item{z}{}

\item{NE}{}
}
\value{
definition of the probability
}
\description{
Define the probability of an event according to the Neighboring cells for a 
protegé cell.
}
