% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_clustering}
\alias{plot_clustering}
\title{Plot clustering}
\usage{
plot_clustering(clust_data, yvar)
}
\arguments{
\item{clust_data}{clustering dataset}

\item{yvar}{character}
}
\description{
Plot clustering
}
