% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_diagram}
\alias{plot_diagram}
\title{Plot diagram over a gradient}
\usage{
plot_diagram(x, ...)
}
\arguments{
\item{data}{}

\item{possible_states}{chr vector specifying the different states}

\item{col_states}{chr vector of the same length of the possible_states vector.
The order of the vector match the order of definition of the possible_states.}
}
\value{
a plot
}
\description{
Plot diagram over a gradient
}
