% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plotnp_gradient}
\alias{plotnp_gradient}
\title{Plot density over a gradient}
\usage{
plotnp_gradient(data, state_var = c("N", "P"), param = c("gamma1", "g"),
  ...)
}
\arguments{
\item{data}{}
}
\value{
a plot
}
\description{
Plot density over a gradient
}
