% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_transition_functions.R
\name{regen}
\alias{regen}
\title{Regeneration transition function}
\usage{
regen(r)
}
\arguments{
\item{r}{rate of regeneration.  1/r defines the expected time before
regeneration.}
}
\value{
definition of the probability
}
\description{
Define the probability to be regenerated for a degraded  cell.
}
