% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{run_2d_gradient}
\alias{run_2d_gradient}
\title{Running a model over a two dimension gradient of parameters}
\usage{
run_2d_gradient(y = "g", x = "gamma1", gradienty = seq(0, 0.2, length.out
  = 5), gradientx = NULL, model_spec = indirect_facilitation_model(),
  time_seq = c(from = 0, to = 100, by = 0.1))
}
\arguments{
\item{y}{A string. The first parameter of the 2d gradient}

\item{x}{A string. The second parameter of the 2d gradient}

\item{gradienty}{A vector. The vector of parameter values used for the
gradient of y.}

\item{gradientx}{A vector. Optional. Same as gradienty but for the second
parameter. If ommited, the function uses the same gradient than for y.}

\item{model_spec}{A SimObj. A model of the object class defined in the
simecol package.}

\item{time_seq}{A vector. A vector containing the timestep values at which
the model will be evaluated.}
}
\value{
A data.frame of tibble data.frame containing the output of the mod#' el
}
\description{
This function runs a SimObj model along a two dimensional gradient of
parameters
}
