% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{run_2d_model}
\alias{run_2d_model}
\title{Run the model by specifying two parameters}
\usage{
run_2d_model(x, y, name_x, name_y, model)
}
\arguments{
\item{x}{First variable.}

\item{y}{Second variable.}

\item{name_x}{A character vector.}

\item{name_y}{A character vector.}
}
\value{
a data.frame
}
\description{
Run the model of a two dimensional gradient.
}
