% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{run_bifurc_model}
\alias{run_bifurc_model}
\title{Bifurcation state diagram}
\usage{
run_bifurc_model(x, scenario = "together", name_x, name_y, model)
}
\arguments{
\item{x}{First variable.}

\item{name_x}{A character vector.}

\item{name_y}{A character vector.}
}
\description{
This function takes as y the vector of initial species densities and as x the
definition of the parameter(s).
}
\details{
The densities of species pairs are defined as
}
\seealso{
init_scenarii
}
