% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{run_scenarii_gradient}
\alias{run_scenarii_gradient}
\title{Run simulation over different initial scenarii}
\usage{
run_scenarii_gradient(
  gradient = list(g = seq(0, 0.3, length.out = 1), b = seq(0, 1, length.out = 1)),
  model_spec = "two_facilitation_model",
  time_seq = c(from = 0, to = 1000, by = 1),
  param = NULL,
  nb_cores = NULL,
  solver_type = NULL,
  scenarii = NULL,
  set_tail = NULL,
  nrep = NULL
)
}
\arguments{
\item{model_spec}{A character string. This is the name of the model function
to use, which calls a SimObj. It can take the values "two_facilitation_model"
or "indirect_facilitation_model"}

\item{time_seq}{A vector. A vector containing the timestep values at which
the model will be evaluated.}

\item{scenarii}{a list of vector. A vector contains initial values of the
model}
}
\description{
A wrapper of run_2d_gradient which takes a list of initial values
}
