% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{run_simecol}
\alias{run_simecol}
\title{Run the model by specifying initial values, parameters and the model}
\usage{
run_simecol(inits, params, model, set_tail = NULL)
}
\arguments{
\item{inits}{a vector of initial values. See simecole::init}

\item{params}{a named vector of length one}

\item{model}{a function containing a odeModel}
}
\value{
a data.frame
}
\description{
Run the model over multidimensional gradient and initial values.
}
