% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{select.scenarii}
\alias{select.scenarii}
\title{Select variables of runs}
\usage{
\method{select}{scenarii}(data, ...)
}
\arguments{
\item{data}{a scenarii object}

\item{...}{variables selected}
}
\value{
a scenarii object
}
\description{
Select variables of runs
}
\details{
https://cran.r-project.org/web/packages/dplyr/vignettes/programming.html
}
\seealso{
dplyr::select
}
