% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_3_states_model.R
\name{steady_state}
\alias{steady_state}
\title{Define a custom solver checking for steady state}
\usage{
steady_state(time, init, func, parms)
}
\description{
Define a custom solver checking for steady state
}
