% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_solver.R
\name{steady_state_3}
\alias{steady_state_3}
\title{Define a custom solver checking for steady state}
\usage{
steady_state_3(time, init, func, parms)
}
\description{
It stops the numerical integration when the sum of the absolute values of the
derivatives become nuls
}
