% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_template.R
\name{steady_state_upca}
\alias{steady_state_upca}
\title{Solver to stop at steady state}
\usage{
steady_state_upca(time, init, func, parms)
}
\arguments{
\item{time}{c(from = 0, to = 100, by = 0.1)}

\item{init}{vector of initial values for variables.}
}
\value{
a list of derivative values.
}
\description{
Solver to stop at steady state
}
