% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_3_states_model.R
\name{three_states_sys}
\alias{three_states_sys}
\title{pair approximation system of ODEs with a nurse, a protege and empty states}
\usage{
three_states_sys(time, init, parms)
}
\arguments{
\item{...}{Nothing}
}
\value{
A list of ODEs
}
\description{
pair approximation system of ODEs with a nurse, a protege and empty states
}
