% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_template.R
\name{upca_model}
\alias{upca_model}
\title{Template to bluid an ode system}
\usage{
upca_model()
}
\arguments{
\item{...}{Nothing}
}
\value{
An ecological model of the class simObj.

  This example is coming from the documentation of the simecol package. See
  \url{http://simecol.r-forge.r-project.org/} for more details.
}
\description{
\code{upca_model} returns the simObj.
}
\details{
This a template from the simecol package. It produces guidelines to easily
compute odes.
}
\examples{
sum(1:10)
sum(1:5, 6:10)
sum(F, F, F, T, T)

sum(.Machine$integer.max, 1L)
sum(.Machine$integer.max, 1)

\dontrun{
sum("a")
}

}
