% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_template.R
\name{upca_ode}
\alias{upca_ode}
\title{ODE definitions}
\usage{
upca_ode(time, init, parms)
}
\arguments{
\item{time}{c(from = 0, to = 100, by = 0.1)}

\item{init}{vector of initial values for variables.}
}
\value{
a list of derivative values.
}
\description{
Describe an ode system
}
