% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{bind_scenar}
\alias{bind_scenar}
\title{Bind scenarii object together}
\usage{
bind_scenar(x, ...)
}
\arguments{
\item{...}{scenarii objects}
}
\description{
Combine a list of scenarii in one
}
\details{
The scenario should describe one of init density values defined in
init_scenarii
}
\seealso{
init_scenarii
}
