% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{check_consistency}
\alias{check_consistency}
\title{Check consistency of cellular automata runs}
\usage{
check_consistency(data, threshold = 0.001)
}
\arguments{
\item{data}{a data.frame}
}
\value{
logical
}
\description{
Check consistency of cellular automata runs
}
\details{
It checks that more than 80% of the simulations give the same
behaviour (e.g. same species win)
}
