% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{compute_states}
\alias{compute_states}
\title{Compute the state result of a range of simulation}
\usage{
compute_states(x, ...)
}
\arguments{
\item{data}{a gradient object}

\item{param}{parameters of the gradient}

\item{type}{}

\item{possibles_states.}{The different states that can take a simulation.
Details in def_state}
}
\description{
Compute the state result of a range of simulation
}
\seealso{
def_state
}
