% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_methods.R
\name{compute_states}
\alias{compute_states}
\title{Compute the state result of a range of simulation}
\usage{
compute_states(x, ...)
}
\arguments{
\item{data}{a gradient object}

\item{param}{parameters of the gradient}

\item{var_to_keep}{variable name that you want to keep, e.g. scenario for
scenarii object (set by default)}

\item{type}{character string. The type of state to be computed, either
"single" or "double" for see bistable areas}

\item{warning_as_desert}{TRUE by default. Convert anormal simulation to
desert (convenient for plotting)}
}
\description{
Compute the state result of a range of simulation
}
\seealso{
def_state
}
