% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_methods.R
\name{run_scenarii_gradient}
\alias{run_scenarii_gradient}
\title{Run simulation over different initial scenarii}
\usage{
run_scenarii_gradient(y = "g", x = "b", gradienty = seq(0, 0.3, by = 0.1),
  gradientx = seq(0, 1, by = 0.1), model_spec = two_facilitation_model(),
  time_seq = c(from = 0, to = 1000, by = 1), param = NULL,
  nb_cores = NULL, solver_type = NULL, scenarii = init_scenarii())
}
\arguments{
\item{y}{A string. The first parameter of the 2d gradient}

\item{x}{A string. The second parameter of the 2d gradient}

\item{gradienty}{A vector. The vector of parameter values used for the
gradient of y.}

\item{gradientx}{A vector. Optional. Same as gradienty but for the second
parameter. If ommited, the function uses the same gradient than for y.}

\item{model_spec}{A SimObj. A model of the object class defined in the
simecol package.}

\item{time_seq}{A vector. A vector containing the timestep values at which
the model will be evaluated.}

\item{scenarii}{a list of vector. A vector contains initial values of the
model}
}
\description{
A wrapper of run_2d_gradient which takes a list of initial values
}
